pyproject.toml
[tool.poetry]
name = "bereal-telegram"
version = "0.1.0"
description = ""
authors = ["Evgenia Kivotova <e.kivotova@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = {version = "20.0a0", allow-prereleases = true}
python-environ = "^0.4.54"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
